+++ src/file_walker.rs
@@ -72,11 +72,13 @@ fn walk_tree_walker(
 
 pub fn walk_directory(
     root: &Path,
+    follow_symlinks: bool,
     toxicity_indicator_calculators: &mut Vec<Box<dyn ToxicityIndicatorCalculator>>,
 ) -> Result<flare::FlareTreeNode, Error> {
     walk_tree_walker(
         WalkBuilder::new(root)
             .add_custom_ignore_filename(".polyglot_code_scanner_ignore")
+            .follow_links(follow_symlinks)
             .sort_by_file_name(|name1, name2| name1.cmp(name2))
             .build(),
         root,
@@ -93,7 +95,15 @@ mod test {
     #[test]
     fn scanning_a_filesystem_builds_a_tree() {
         let root = Path::new("./tests/data/simple/");
-        let tree = walk_directory(root, &mut Vec::new()).unwrap();
+        let tree = walk_directory(root, false, &mut Vec::new()).unwrap();
+
+        assert_eq_json_file(&tree, "./tests/expected/simple_files.json")
+    }
+
+    #[test]
+    fn scanning_a_filesystem_can_follow_symlinks() {
+        let root = Path::new("./tests/data/simple_linked/");
+        let tree = walk_directory(root, true, &mut Vec::new()).unwrap();
 
         assert_eq_json_file(&tree, "./tests/expected/simple_files.json")
     }
@@ -146,7 +156,7 @@ mod test {
         let calculators: &mut Vec<Box<dyn ToxicityIndicatorCalculator>> =
             &mut vec![Box::new(simple_tic), Box::new(self_naming_tic)];
 
-        let tree = walk_directory(root, calculators).unwrap();
+        let tree = walk_directory(root, false, calculators).unwrap();
 
         assert_eq_json_file(&tree, "./tests/expected/simple_files_with_data.json");
     }
@@ -177,7 +187,7 @@ mod test {
         let tic = MutableTIC { count: 0 };
         let calculators: &mut Vec<Box<dyn ToxicityIndicatorCalculator>> = &mut vec![Box::new(tic)];
 
-        let tree = walk_directory(root, calculators).unwrap();
+        let tree = walk_directory(root, false, calculators).unwrap();
 
         assert_eq_json_file(&tree, "./tests/expected/simple_files_with_counts.json");
     }