+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
+
+### Changed
+
+* Added an option to follow symlinks to fix issue #1
+
 ## [0.3.0] - 2021-04-05
 
 ### Changed